Fix: 아카이브 기록하기에서 별점 기본 5개, 등록시 알람팝업 하나, 기록후 마이아카이브로 넘어가기
