+++ components/Organisms/Popup/PopupRouter.tsx
@@ -1,12 +1,11 @@
 import { useRecoilValue, useSetRecoilState } from 'recoil';
 
-import AlertArchiveAskPopup from 'components/Organisms/Popup/AlertArchiveAskPopup';
 import AlertCancelConfirmPopup from 'components/Organisms/Popup/AlertCancelConfirmPopup';
 import AlertSuccessPopup from 'components/Organisms/Popup/AlertConfirmPopup';
 import AlertLoginConfirmationPopup from 'components/Organisms/Popup/AlertLoginConfirmationPopup';
+import AlertMoveToMyArchivePopup from 'components/Organisms/Popup/AlertMoveToMyArchivePopup';
 import ArchivePictureOverflowPopup from 'components/Organisms/Popup/ArchivePictureOverflowPopup';
 import ArchiveWirteConfirmPopup from 'components/Organisms/Popup/ArchiveWirteConfirmPopup';
-import ArchiveWirteErrorPopup from 'components/Organisms/Popup/ArchiveWirteErrorPopup';
 import MyArchiveDetailPopup from 'components/Organisms/Popup/MyArchiveDetailPopup';
 import { POPUP_NAME } from 'constants/popupName';
 import { PopupNameState } from 'states';
@@ -31,12 +30,10 @@ const PopupRouter = () => {
       return <AlertLoginConfirmationPopup />;
     case POPUP_NAME.POPUP_ARCHIVE_DETAIL:
       return <MyArchiveDetailPopup />;
-    case POPUP_NAME.ALERT_ARCHIVE_ASK:
-      return <AlertArchiveAskPopup />;
+    case POPUP_NAME.ALERT_MOVE_MyARCHIVE_PAGE:
+      return <AlertMoveToMyArchivePopup />;
     case POPUP_NAME.ARCHIVE_WRITE_CONFIRM:
       return <ArchiveWirteConfirmPopup />;
-    case POPUP_NAME.ARCHIVE_WRITE_ERROR:
-      return <ArchiveWirteErrorPopup />;
     case POPUP_NAME.OVERFLOW_PICTURE:
       return <ArchivePictureOverflowPopup />;
     default:
